Move getRelease into root cmd and refactor.
